+++ src/backend/efiling-api/src/test/java/ca/bc/gov/open/jag/efilingapi/TestHelpers.java
@@ -266,8 +266,8 @@ public class TestHelpers {
         ReviewDocument reviewDocumentOne = new ReviewDocument();
         reviewDocumentOne.setDocumentId(DOCUMENT_ID_ONE);
         reviewDocumentOne.setFileName(NAME);
-        reviewDocumentOne.setDocumentTypeCd(InitialDocument.TypeEnum.AAB.getValue());
         reviewDocumentOne.setDocumentType(DESCRIPTION);
+        reviewDocumentOne.setDocumentTypeCd(DocumentProperties.TypeEnum.AAB.getValue());
         reviewDocumentOne.setStatus(STATUS);
         reviewDocumentOne.setStatusCode(STATUS_CODE);
         reviewDocumentOne.setStatusDate(DateTime.parse("2020-05-05T00:00:00.000-07:00"));
@@ -276,8 +276,8 @@ public class TestHelpers {
         ReviewDocument reviewDocumentTwo = new ReviewDocument();
         reviewDocumentTwo.setDocumentId(DOCUMENT_ID_TWO);
         reviewDocumentTwo.setFileName(NAME);
-        reviewDocumentTwo.setDocumentTypeCd(InitialDocument.TypeEnum.AAB.getValue());
         reviewDocumentTwo.setDocumentType(DESCRIPTION);
+        reviewDocumentTwo.setDocumentTypeCd(DocumentProperties.TypeEnum.AAB.getValue());
         reviewDocumentTwo.setStatus(STATUS);
         reviewDocumentTwo.setStatusCode(STATUS_CODE);
         reviewDocumentTwo.setStatusDate(DateTime.parse("2020-05-05T00:00:00.000-07:00"));
